+++ apps/browser/src/popup/settings/folders.component.ts
@@ -1,7 +1,7 @@
 import { Component, OnInit } from "@angular/core";
 import { Router } from "@angular/router";
 
-import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
+import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
 import { FolderView } from "@bitwarden/common/models/view/folderView";
 
 @Component({
@@ -11,10 +11,10 @@ import { FolderView } from "@bitwarden/common/models/view/folderView";
 export class FoldersComponent implements OnInit {
   folders: FolderView[];
 
-  constructor(private folderStateService: FolderStateService, private router: Router) {}
+  constructor(private folderService: FolderService, private router: Router) {}
 
   async ngOnInit() {
-    this.folders = await this.folderStateService.getAllDecrypted();
+    this.folders = await this.folderService.getAllDecrypted();
     // Remove "No Folder"
     if (this.folders.length > 0) {
       this.folders = this.folders.slice(0, this.folders.length - 1);
