+++ apps/browser/src/popup/vault/add-edit.component.ts
@@ -8,7 +8,7 @@ import { AuditService } from "@bitwarden/common/abstractions/audit.service";
 import { CipherService } from "@bitwarden/common/abstractions/cipher.service";
 import { CollectionService } from "@bitwarden/common/abstractions/collection.service";
 import { EventService } from "@bitwarden/common/abstractions/event.service";
-import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
+import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/abstractions/log.service";
 import { MessagingService } from "@bitwarden/common/abstractions/messaging.service";
@@ -35,7 +35,7 @@ export class AddEditComponent extends BaseAddEditComponent {
 
   constructor(
     cipherService: CipherService,
-    folderStateService: FolderStateService,
+    folderService: FolderService,
     i18nService: I18nService,
     platformUtilsService: PlatformUtilsService,
     auditService: AuditService,
@@ -54,7 +54,7 @@ export class AddEditComponent extends BaseAddEditComponent {
   ) {
     super(
       cipherService,
-      folderStateService,
+      folderService,
       i18nService,
       platformUtilsService,
       auditService,
