+++ apps/cli/src/commands/create.command.ts
@@ -4,8 +4,8 @@ import * as path from "path";
 import { ApiService } from "@bitwarden/common/abstractions/api.service";
 import { CipherService } from "@bitwarden/common/abstractions/cipher.service";
 import { CryptoService } from "@bitwarden/common/abstractions/crypto.service";
-import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
-import { FolderServiceAbstraction } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
+import { FolderApiServiceAbstraction } from "@bitwarden/common/abstractions/folder/folder-api.service.abstraction";
+import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
 import { StateService } from "@bitwarden/common/abstractions/state.service";
 import { Utils } from "@bitwarden/common/misc/utils";
 import { CipherExport } from "@bitwarden/common/models/export/cipherExport";
@@ -24,11 +24,11 @@ import { CliUtils } from "../utils";
 export class CreateCommand {
   constructor(
     private cipherService: CipherService,
-    private folderStateService: FolderStateService,
+    private folderService: FolderService,
     private stateService: StateService,
     private cryptoService: CryptoService,
     private apiService: ApiService,
-    private folderService: FolderServiceAbstraction
+    private folderApiService: FolderApiServiceAbstraction
   ) {}
 
   async run(
@@ -148,10 +148,10 @@ export class CreateCommand {
   }
 
   private async createFolder(req: FolderExport) {
-    const folder = await this.folderStateService.encrypt(FolderExport.toView(req));
+    const folder = await this.folderService.encrypt(FolderExport.toView(req));
     try {
-      await this.folderService.save(folder);
-      const newFolder = await this.folderStateService.get(folder.id);
+      await this.folderApiService.save(folder);
+      const newFolder = await this.folderService.get(folder.id);
       const decFolder = await newFolder.decrypt();
       const res = new FolderResponse(decFolder);
       return Response.success(res);
