+++ apps/cli/src/commands/delete.command.ts
@@ -1,7 +1,7 @@
 import { ApiService } from "@bitwarden/common/abstractions/api.service";
 import { CipherService } from "@bitwarden/common/abstractions/cipher.service";
-import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
-import { FolderServiceAbstraction } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
+import { FolderApiServiceAbstraction } from "@bitwarden/common/abstractions/folder/folder-api.service.abstraction";
+import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
 import { StateService } from "@bitwarden/common/abstractions/state.service";
 import { Utils } from "@bitwarden/common/misc/utils";
 import { Response } from "@bitwarden/node/cli/models/response";
@@ -11,10 +11,10 @@ import { CliUtils } from "src/utils";
 export class DeleteCommand {
   constructor(
     private cipherService: CipherService,
-    private folderStateService: FolderStateService,
+    private folderService: FolderService,
     private stateService: StateService,
     private apiService: ApiService,
-    private folderService: FolderServiceAbstraction
+    private folderApiService: FolderApiServiceAbstraction
   ) {}
 
   async run(object: string, id: string, cmdOptions: Record<string, any>): Promise<Response> {
@@ -88,13 +88,13 @@ export class DeleteCommand {
   }
 
   private async deleteFolder(id: string) {
-    const folder = await this.folderStateService.get(id);
+    const folder = await this.folderService.get(id);
     if (folder == null) {
       return Response.notFound();
     }
 
     try {
-      await this.folderService.delete(id);
+      await this.folderApiService.delete(id);
       return Response.success();
     } catch (e) {
       return Response.error(e);
