+++ apps/cli/src/send.program.ts
@@ -144,7 +144,7 @@ export class SendProgram extends Program {
       .action(async (object) => {
         const cmd = new GetCommand(
           this.main.cipherService,
-          this.main.folderStateService,
+          this.main.folderService,
           this.main.collectionService,
           this.main.totpService,
           this.main.auditService,
