+++ apps/desktop/src/app/vault/folder-add-edit.component.ts
@@ -1,8 +1,8 @@
 import { Component } from "@angular/core";
 
 import { FolderAddEditComponent as BaseFolderAddEditComponent } from "@bitwarden/angular/components/folder-add-edit.component";
-import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
-import { FolderServiceAbstraction } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
+import { FolderApiServiceAbstraction } from "@bitwarden/common/abstractions/folder/folder-api.service.abstraction";
+import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/abstractions/log.service";
 import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUtils.service";
@@ -13,12 +13,12 @@ import { PlatformUtilsService } from "@bitwarden/common/abstractions/platformUti
 })
 export class FolderAddEditComponent extends BaseFolderAddEditComponent {
   constructor(
-    folderStateService: FolderStateService,
-    folderService: FolderServiceAbstraction,
+    folderService: FolderService,
+    folderApiService: FolderApiServiceAbstraction,
     i18nService: I18nService,
     platformUtilsService: PlatformUtilsService,
     logService: LogService
   ) {
-    super(folderStateService, folderService, i18nService, platformUtilsService, logService);
+    super(folderService, folderApiService, i18nService, platformUtilsService, logService);
   }
 }
