+++ libs/angular/src/components/add-edit.component.ts
@@ -4,7 +4,7 @@ import { AuditService } from "@bitwarden/common/abstractions/audit.service";
 import { CipherService } from "@bitwarden/common/abstractions/cipher.service";
 import { CollectionService } from "@bitwarden/common/abstractions/collection.service";
 import { EventService } from "@bitwarden/common/abstractions/event.service";
-import { FolderStateService } from "@bitwarden/common/abstractions/folder/folder-state.service.abstraction";
+import { FolderService } from "@bitwarden/common/abstractions/folder/folder.service.abstraction";
 import { I18nService } from "@bitwarden/common/abstractions/i18n.service";
 import { LogService } from "@bitwarden/common/abstractions/log.service";
 import { MessagingService } from "@bitwarden/common/abstractions/messaging.service";
@@ -79,7 +79,7 @@ export class AddEditComponent implements OnInit {
 
   constructor(
     protected cipherService: CipherService,
-    protected folderStateService: FolderStateService,
+    protected folderService: FolderService,
     protected i18nService: I18nService,
     protected platformUtilsService: PlatformUtilsService,
     protected auditService: AuditService,
@@ -243,7 +243,7 @@ export class AddEditComponent implements OnInit {
       }
     }
 
-    this.folders = await this.folderStateService.getAllDecrypted();
+    this.folders = await this.folderService.getAllDecrypted();
 
     if (this.editMode && this.previousCipherId !== this.cipherId) {
       this.eventService.collect(EventType.Cipher_ClientViewed, this.cipherId);
